traits timeout type changed to float from int

+++ src/services/AmoTrait.php
@@ -5,6 +5,6 @@ namespace crmpbx\commutator\services;
 trait AmoTrait
 {
     public string $amoServiceAddress = '';
-    public int $amoServiceTimeout = 1;
+    public float $amoServiceTimeout = 1;
     public string $amoTestServiceAddress = '';
 }
