+++ src/services/ExtensionTrait.php
@@ -5,5 +5,5 @@ namespace crmpbx\commutator\services;
 trait ExtensionTrait
 {
     public string $extensionServiceAddress = '';
-    public int $extensionServiceTimeout = 1;
+    public float $extensionServiceTimeout = 1;
 }
